refactor: refactor and introduce XedExtensionPoint annotation
- Introduce `@XedExtensionPoint` annotation to identify extension-facing APIs.
- Migrate `FileActionUtils` to `FileActionProvider` object.
- Implement registration and unregistration methods with duplicate checks across `BuiltinFileType`, `CommandProvider`, `LspRegistry`, `IntelligentFeature`, `TabRegistry`, and `RunnerManager`.
- Apply `@XedExtensionPoint` to relevant registry management functions.
