+++ core/main/src/main/java/com/rk/extension/XedExtensionPoint.kt
@@ -1,0 +1,5 @@
+package com.rk.extension
+
+@Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION, AnnotationTarget.PROPERTY)
+@Retention(AnnotationRetention.SOURCE)
+annotation class XedExtensionPoint
